+++ src/components/Grid/Container.js
@@ -3,15 +3,21 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 
 const Container = (props) => {
-  const { children, className, flavor, ...rest } = props;
+  const { children, className, flavor, size, ...rest } = props;
   const flavorClasses = Array.isArray(flavor) ? flavor.map(f => `slds-container_${f}`) : `slds-container_${flavor}`;
-  return (<div {...rest} className={cx([className, flavorClasses])}>{children}</div>);
+  const sldsClasses = [
+    className,
+    flavorClasses,
+    { [`slds-container_${size}`]: !!size }
+  ];
+  return (<div {...rest} className={cx(sldsClasses)}>{children}</div>);
 };
 
 Container.defaultProps = {
   children: null,
   className: null,
-  flavor: ['left', 'small']
+  flavor: ['left'],
+  size: 'small',
 };
 
 Container.propTypes = {
@@ -24,28 +30,28 @@ Container.propTypes = {
    */
   className: PropTypes.string,
   /**
-   * flavor: string or array of strings. Flavors: small, medium, large, x-large,
-   * fluid, left, center, right
+   * flavor: string or array of strings. Flavors: fluid, left, center, right
    */
   flavor: PropTypes.oneOfType([PropTypes.oneOf([
-    'small',
-    'medium',
-    'large',
-    'x-large',
     'fluid',
     'left',
     'center',
     'right',
   ]), PropTypes.arrayOf(PropTypes.oneOf([
-    'small',
-    'medium',
-    'large',
-    'x-large',
     'fluid',
     'left',
     'center',
     'right',
   ]))]),
+  /*
+   * sizes: small, medium, large, x-large
+   */
+  size: PropTypes.oneOf([
+    'small',
+    'medium',
+    'large',
+    'x-large'
+  ])
 };
 
 export default Container;